# backend/docs/swagger.yaml
definitions:
...
  dto.DisciplineResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      external_id:
        type: string
      id:
        type: string
      name:
        type: string
      source:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/me/disciplines/student:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DisciplineResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Мои дисциплины (как студент)
      tags:
      - disciplines
  /api/me/disciplines/teacher:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DisciplineResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Мои дисциплины (как преподаватель)
      tags:
      - disciplines
...
  /api/users/{id}/disciplines/student:
    get:
      parameters:
      - description: UUID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DisciplineResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Дисциплины пользователя (как студент)
      tags:
      - disciplines
  /api/users/{id}/disciplines/teacher:
    get:
      parameters:
      - description: UUID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DisciplineResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Дисциплины пользователя (как преподаватель)
      tags:
      - disciplines
swagger: "2.0"